% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{assignValues}
\alias{assignValues}
\title{Assign values of character strings}
\usage{
assignValues(obj, option)
}
\arguments{
\item{obj}{MRexperiment object.}

\item{option}{Option value, either character or numeric.}
}
\value{
value
}
\description{
Convert values set for aggregateAtDepth or maxDepth to integers.
}
\seealso{
\code{\link{metavizControl}} \code{\link{startMetaviz}}
}

