% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InnerNodesMetavizGraph-class.R
\docType{methods}
\name{buildInnerNodesMetavizGraph}
\alias{buildInnerNodesMetavizGraph}
\alias{buildInnerNodesMetavizGraph,MRexperiment-method}
\title{Build a MetavizTree object from another object}
\usage{
buildInnerNodesMetavizGraph(object, ...)

\S4method{buildInnerNodesMetavizGraph}{MRexperiment}(object, feature_order, ...)
}
\arguments{
\item{object}{The object from which taxonomy data is extracted}

\item{...}{Additional arguments}

\item{feature_order}{Ordering of leaves (features) in taxonomy tree}
}
\value{
a \code{\link{InnerNodesMetavizGraph}} object
}
\description{
Build a MetavizTree object from another object
}
\section{Methods (by class)}{
\itemize{
\item \code{MRexperiment}: Build graph from a \code{\link[metagenomeSeq]{MRexperiment-class}} object
}}

