% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetavizTree-class.R
\docType{methods}
\name{buildMetavizTree}
\alias{buildMetavizTree}
\alias{buildMetavizTree,MRexperiment-method}
\title{Build a MetavizTree object from another object}
\usage{
buildMetavizTree(object, ...)

\S4method{buildMetavizTree}{MRexperiment}(object, feature_order, ...)
}
\arguments{
\item{object}{The object from which taxonomy data is extracted}

\item{...}{Additional arguments}

\item{feature_order}{Ordering of leaves (features) in taxonomy tree}
}
\value{
a \code{\link{MetavizTree}} object
}
\description{
Build a MetavizTree object from another object
}
\section{Methods (by class)}{
\itemize{
\item \code{MRexperiment}: Build tree from a \code{\link[metagenomeSeq]{MRexperiment-class}} object
}}

