% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{checkValues}
\alias{checkValues}
\title{Check appropriate values}
\usage{
checkValues(obj, option)
}
\arguments{
\item{obj}{MRexperiment object.}

\item{option}{Option value.}
}
\description{
Ensure options sent to addMeasurements are reasonable.
}
\seealso{
\code{\link{metavizControl}} \code{\link{startMetaviz}}
}

