% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizMetagenomicsDataTimeSeries-class.R
\docType{class}
\name{EpivizMetagenomicsDataTimeSeries-class}
\alias{EpivizMetagenomicsDataTimeSeries-class}
\alias{EpivizMetagenomicsDataTimeSeries}
\title{Data container for MRexperiment objects}
\description{
Used to serve metagenomic data (used in e.g., icicle plots and heatmaps). Wraps
\code{\link[metagenomeSeq]{MRexperiment-class}} objects.
}

\examples{

library(metagenomeSeq)
data(mouseData)
obj <- metavizr:::EpivizMetagenomicsData$new(mouseData, feature_order = colnames(fData(mouseData)))

}
