% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizMetagenomicsMeasurement-class.R
\docType{class}
\name{EpivizMetagenomicsMeasurement-class}
\alias{EpivizMetagenomicsMeasurement}
\alias{EpivizMetagenomicsMeasurement-class}
\title{Metagenomic Measurement object}
\description{
Metagenomic Measurement object
}

