% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateSelection}
\alias{generateSelection}
\title{Method to select and set aggregation type to nodes in FacetZoom}
\usage{
generateSelection(feature_names, aggregation_level, selection_type,
  feature_order = NULL)
}
\arguments{
\item{feature_names}{Selected Features}

\item{aggregation_level}{Level in the hierarchy}

\item{selection_type}{Expanded, aggregated, or removed}

\item{feature_order}{Order of features at that level}
}
\value{
A selection object for a metavizControl object to accept
}
\description{
Method to select and set aggregation type to nodes in FacetZoom
}
