% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InnerNodesEpivizMetagenomicsData-class.R
\docType{class}
\name{InnerNodesEpivizMetagenomicsData-class}
\alias{InnerNodesEpivizMetagenomicsData-class}
\alias{InnerNodesEpivizMetagenomicsData}
\title{Data container for MRexperiment objects}
\description{
Used to serve metagenomic data (used in e.g., icicle plots and heatmaps). Wraps
\code{\link[metagenomeSeq]{MRexperiment-class}} objects.
}
\section{Methods}{

\describe{
\item{\code{getAlphaDiversity(measurements = NULL)}}{Compute alpha diversity using vegan for the given samples

\describe{
\item{measurements}{Samples to compute alpha diversity}
\item{start}{Start of feature range to query }
\item{end}{End of feature range to query}
}}

\item{\code{getHierarchy(nodeId = NULL)}}{Retrieve feature hierarchy information for subtree with specified root

\describe{
  \item{nodeId}{Feature identifier with level info}
}}

\item{\code{get_measurements()}}{Get all annotation info for all samples

\describe{
  \item{chart_id_or_object}{An object of class \code{EpivizChart} or an id for
    a chart loaded to the epiviz app.}
}}

\item{\code{getPCA(measurements = NULL)}}{Compute PCA over all features for given samples

\describe{
\item{measurements}{Samples to compute PCA over}
\item{start}{Start of feature range to query }
\item{end}{End of feature range to query}
}}

\item{\code{getRows(measurements = NULL, start = 1, end = 1000, selectedLevels = 3,
  selections = NULL)}}{Return the sample annotation and features within the specified range and level for a given sample and features
    
    \describe{
      \item{measurements}{Samples to retrieve for}
      \item{start}{Start of feature range to query}
      \item{end}{End of feature range to query}
      \item{selections}{Node-id and selectionType pairs}
      \item{selectedLevels}{Current aggregation level}

    }
    }

\item{\code{getValues(measurements = NULL, start = 1, end = 1000,
  selectedLevels = 3, selections = NULL)}}{Return the counts for a sample within the specified range

\describe{
\item{measurements}{Samples to get counts for}
\item{start}{Start of feature range to query}
\item{end}{End of feature range to query}
\item{selections}{Node-id and selectionType pairs}
\item{selectedLevels}{Current aggregation level}

}}

\item{\code{propagateHierarchyChanges(selection = NULL, order = NULL,
  selectedLevels = NULL, request_with_labels = FALSE)}}{Update internal state for hierarchy

\describe{
  \item{selection}{Node-id and selectionType pairs}
  \item{order}{Ordering of features}
  \item{selectedLevels}{Current aggregation level}
  \item{request_with_labels}{For handling requests using fData entries from MRexperiment}
}}

\item{\code{searchTaxonomy(query = NULL, max_results = 15)}}{Return list of features matching a text-based query

\describe{
\item{query}{String of feature for which to search}
\item{max_results}{Maximum results to return}

}}
}}

\examples{

\dontrun{
library(curatedMetagenomicData)
zeller.eset = ZellerG_2014.metaphlan_bugs_list.stool()
zeller_MR <- ExpressionSet2MRexperiment(zeller.eset)
feature_order <- colnames(fData(zeller_MR))
sampleId<- "CCIS98482370ST-3-0"
mObj <- metavizr:::InnerNodesEpivizMetagenomicsData$new(zeller_MR, feature_order = feature_order)
}

}
