% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InnerNodesMetavizGraph-class.R
\docType{class}
\name{InnerNodesMetavizGraph-class}
\alias{InnerNodesMetavizGraph-class}
\alias{InnerNodesMetavizGraph}
\title{Graph implementation to query hierarchical feature data}
\description{
Used to manage aggregation and range queries from the Metaviz app UI.
}

