% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{metavizControl}
\alias{metavizControl}
\title{metavizr settings}
\usage{
metavizControl(aggregateAtDepth = 3, aggregateFun = function(x) log2(1 +
  colSums(x)), valuesAnnotationFuns = NULL, maxDepth = 4, maxHistory = 3,
  maxValue = NULL, minValue = NULL, title = "", n = 10000,
  rankFun = sd, norm = TRUE, log = FALSE)
}
\arguments{
\item{aggregateAtDepth}{Level of the tree to aggregate counts at by default.}

\item{aggregateFun}{Function to aggregate counts by at the aggregateAtDepth level.}

\item{valuesAnnotationFuns}{Function for error bars.}

\item{maxDepth}{Level of the tree to display by default in icicle view.}

\item{maxHistory}{Value for caching.}

\item{maxValue}{Maximum value to display.}

\item{minValue}{Minimum value to display.}

\item{title}{title.}

\item{n}{Number of OTUs to include in ranking.}

\item{rankFun}{Ranking function - single vector function.}

\item{norm}{Normalize MRexperiment object.}

\item{log}{Log tranformation of MRexperiment object.}
}
\value{
List of setting parameters.
}
\description{
Default settings for the various plotting functions in metavizr.
}
\examples{
settings = metavizControl()

}
\seealso{
\code{\link{metavizControl}} \code{\link{metavizRank}} \code{\link{metavizOptimize}}
}

