% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetavizGraph-class.R
\docType{class}
\name{MetavizGraph-class}
\alias{MetavizGraph-class}
\alias{MetavizGraph}
\title{Taxonomy tree structure wrapper}
\description{
Used to manage tree-cut queries from the
Metaviz app UI.
}

