% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetavizGraphInnerNodes-class.R
\docType{class}
\name{MetavizGraphInnerNodes-class}
\alias{MetavizGraphInnerNodes-class}
\alias{MetavizGraphInnerNodes}
\title{Graph implementation to query hierarchical feature data}
\description{
Used to manage aggregation and range queries from the Metaviz app UI.
}

