% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{metavizHeatmap}
\alias{metavizHeatmap}
\title{Heatmap}
\usage{
metavizHeatmap(obj, mgr, control = metavizControl(title = "heatmap"),
  samples = NULL)
}
\arguments{
\item{obj}{MRexperiment object or EpivizMetagenomicsData object.
If EpivizMetagenomicsData object then control is ignored.}

\item{mgr}{manager of a metaviz session.}

\item{control}{List of options passed through `metavizControl`.}

\item{samples}{Index vector of samples to include in plot.}
}
\value{
EpivizMetagenomicsData class object
}
\description{
Produces a heatmap for the experiment.
}
\examples{
#todo
}
\seealso{
\code{\link{metavizControl}} \code{\link{metaviztree}} \code{\link{metavizRank}} \code{\link{metavizOptimize}}
}

