% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{metavizMDS}
\alias{metavizMDS}
\title{MDS (multiple dimensional scaling) or PCA}
\usage{
metavizMDS(obj, mgr, usePCA = FALSE, useDist = TRUE, cord = c(1, 2),
  distFun = stats::dist, distMethod = "euclidian", tran = FALSE,
  control = metavizControl(title = "MDS plot"))
}
\arguments{
\item{obj}{MRexperiment object.}

\item{mgr}{manager of a metaviz session.}

\item{usePCA}{TRUE/FALSE whether to use PCA or MDS coordinates (TRUE isPCA).}

\item{useDist}{TRUE/FALSE whether to calculate distance matrix.}

\item{cord}{Which coordinates to compare/display.}

\item{distFun}{Distance function - default is stats::dist.}

\item{distMethod}{If useDist==TRUE, what method to calculate distances.}

\item{tran}{Transpose the matrix.}

\item{control}{List of options passed through `metavizControl`.}
}
\value{
EpivizMetagenomicsData class object
}
\description{
Produces a PCA or PCoA scatterplot of an experiment.
}
\examples{
data(mouseData)
# metavizOrd(mouseData,mgr)
}
\seealso{
\code{\link{metavizControl}} \code{\link{metavizRank}} \code{\link{cmdscale}} \code{\link{prcomp}}
}

