% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{metavizRank}
\alias{metavizRank}
\title{Indices of ranked features}
\usage{
metavizRank(obj, control = metavizControl(log = TRUE))
}
\arguments{
\item{obj}{MRexperiment object.}

\item{control}{List of options passed through `metavizControl`.}
}
\value{
value
}
\description{
Calculate top n features at the leaves. Returns indices for the rankings.
}
\examples{
# 10 highest mean OTUs
indices = metavizRank(msd16s[,1:10],control=metavizControl(n=10,rankFun=mean,log=TRUE))
# 5 highest sd OTUs
indices = metavizRank(msd16s[,1:10],control=metavizControl(n=5, rankFun=sd,log=TRUE))
# 20 highest max OTUs
indices = metavizRank(msd16s[,1:10],control=metavizControl(n=20,rankFun=max,log=TRUE))
}
\seealso{
\code{\link{metavizControl}} \code{\link{metavizRank}}
}

