% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetavizTree-class.R
\docType{class}
\name{MetavizTree-class}
\alias{MetavizTree-class}
\alias{MetavizTree}
\title{Taxonomy tree structure wrapper that is used to manage tree-cut queries from the
Metaviz app UI.}
\description{
Taxonomy tree structure wrapper that is used to manage tree-cut queries from the
Metaviz app UI.
}

