% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{metavizTree}
\alias{metavizTree}
\title{Icicle}
\usage{
metavizTree(datasource, mgr)
}
\arguments{
\item{datasource}{An EpivizMetagenomicsData class object (i.e. output from one of the metavizPlot functions).}

\item{mgr}{manager of a metaviz session.}
}
\value{
Icicle plot
}
\description{
Produces an icicle for the tree of an EpivizMetagenomicsData class object
}
\examples{
# assuming datasource, and mgr
# metavizTree(datasource,mgr)
}
\seealso{
\code{\link{startMetaviz}}
}

