% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-methods.R
\docType{methods}
\name{register,MRexperiment-method}
\alias{register,MRexperiment-method}
\title{Generic method to register data to the epiviz data server}
\usage{
\S4method{register}{MRexperiment}(object, type = "LeafCounts",
  columns = NULL, ...)
}
\arguments{
\item{object}{The object to register to data server}

\item{type}{leafCounts, if data objects has counts at leaf level
or innerNodeCounts, if data object has counts at inner nodes}

\item{columns}{Name of columns containing data to register}

\item{...}{Additonal arguments passed to object constructors}
}
\value{
An \code{\link{EpivizMetagenomicsData-class}} object
}
\description{
Generic method to register data to the epiviz data server
}
