% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-methods.R
\name{register,phyloseq-method}
\alias{register,phyloseq-method}
\title{Generic method to register data to the epiviz data server}
\usage{
\S4method{register}{phyloseq}(object, type = "LeafCounts", ...)
}
\arguments{
\item{object}{The object to register to data server}

\item{type}{leafCounts, if data objects has counts at leaf level
or innerNodeCounts, if data object has counts at inner nodes}

\item{...}{Additonal arguments passed to object constructors}
}
\value{
An \code{\link{phyloseq-class}} object
}
\description{
Generic method to register data to the epiviz data server
}
