% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-methods.R
\name{register,TreeSummarizedExperiment-method}
\alias{register,TreeSummarizedExperiment-method}
\title{Generic method to register data to the epiviz data server}
\usage{
\S4method{register}{TreeSummarizedExperiment}(object, tree = "row", columns = NULL, ...)
}
\arguments{
\item{object}{The object to register to data server}

\item{columns}{Name of columns containing data to register}

\item{...}{Additonal arguments passed to object constructors}
}
\value{
An \code{\link{EpivizMetagenomicsData-class}} object
}
\description{
Generic method to register data to the epiviz data server
}
