% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-methods.R
\docType{methods}
\name{register,MRexperiment-method}
\alias{register,MRexperiment-method}
\title{Generic method to register data to the epiviz data server}
\usage{
\S4method{register}{MRexperiment}(object, columns = NULL, ...)
}
\arguments{
\item{object}{The object to register to data server}

\item{...}{Additonal arguments passed to object constructors}

\item{object}{Which object to register}
}
\value{
Object inheriting from \code{\link{EpivizData}} class
}
\description{
Generic method to register data to the epiviz data server
}
\section{Methods (by class)}{
\itemize{
\item \code{MRexperiment}: Register a \code{\link{MRexperiment}} object
}}

