% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replaceNAFeatures}
\alias{replaceNAFeatures}
\title{Method to replace NA or null feature labels with Not_Annotated_hierarchy-level}
\usage{
replaceNAFeatures(replacing_na_obj_fData, feature_order)
}
\arguments{
\item{replacing_na_obj_fData}{fData from MRexperiment object to replace NA or null}

\item{feature_order}{Order of features}
}
\value{
hierarchy with NA or null feature labels replaced
}
\description{
Method to replace NA or null feature labels with Not_Annotated_hierarchy-level
}
\examples{
library(msd16s)
feature_order <- colnames(fData(msd16s))[1:8]
replaceNAFeatures(fData(msd16s)[,feature_order],feature_order)
}
