% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startMetaviz.R
\name{setMetavizStandalone}
\alias{setMetavizStandalone}
\title{set metaviz app standalone settings}
\usage{
setMetavizStandalone(url = "https://github.com/epiviz/epiviz.git",
  branch = "metaviz-4.1", local_path = NULL, non_interactive = FALSE)
}
\arguments{
\item{url}{(character) github url to use. defaults to (\url{"https://github.com/epiviz/epiviz.git"}).}

\item{branch}{(character) branch on the github repository. defaults to (master).}

\item{local_path}{(character) if you already have a local instance of metaviz and would like to run standalone use this.}

\item{non_interactive}{(logical) don't download repo, used for testing purposes.}
}
\value{
path to the metaviz app git repository
}
\description{
set metaviz app standalone settings
}
\examples{

\dontrun{
# see package vignette for example usage
setMetavizStandalone()
}

}
