% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startMetaviz.R
\name{startMetaviz}
\alias{startMetaviz}
\title{Start metaviz app and create \code{\link[metavizr]{MetavizApp}} object to manage connection.}
\usage{
startMetaviz(
  host = "http://metaviz.cbcb.umd.edu",
  register_function = .register_all_metaviz_things,
  ...
)
}
\arguments{
\item{host}{(character) host address to launch.}

\item{register_function}{(function) function used to register actions and charts on the metaviz app.}

\item{...}{additional parameters passed to \code{\link[epivizr]{startEpiviz}}.}
}
\value{
An object of class \code{\link[metavizr]{MetavizApp}}
}
\description{
Start metaviz app and create \code{\link[metavizr]{MetavizApp}} object to manage connection.
}
\examples{
# see package vignette for example usage
app <- startMetaviz(non_interactive=TRUE, open_browser=FALSE)
app$stop_app()

}
\seealso{
\code{\link[metavizr]{MetavizApp}}
}
