% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startMetaviz.R
\name{startMetavizStandalone}
\alias{startMetavizStandalone}
\title{Start metaviz app in standalone (locally) and create \code{\link[metavizr]{MetavizApp}} object to manage connection.}
\usage{
startMetavizStandalone(branch = "metaviz-4.1",
  register_function = .register_all_metaviz_things, ...)
}
\arguments{
\item{branch}{(character) branch to pull from metaviz github repo to run standalone.}

\item{register_function}{(function) function used to register actions and charts on the metaviz app.}

\item{...}{additional parameters passed to \code{\link[epivizrStandalone]{startStandalone}}.}
}
\value{
An object of class \code{\link{MetavizApp}}
}
\description{
Start metaviz app in standalone (locally) and create \code{\link[metavizr]{MetavizApp}} object to manage connection.
}
\examples{
# see package vignette for example usage
app <- startMetaviz(non_interactive=TRUE, open_browser=TRUE)
app$stop_app()

}

