% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizMetagenomicsData-class.R
\docType{class}
\name{EpivizMetagenomicsData-class}
\alias{EpivizMetagenomicsData}
\alias{EpivizMetagenomicsData-class}
\title{Data container for MRexperiment objects}
\description{
Used to serve metagenomic data (used in e.g., icicle plots and heatmaps). Wraps
\code{\link{MRexperiment}} objects.
}
\section{Methods}{

\describe{
\item{\code{get_default_chart_type()}}{Get name of default chart type for this data type}

\item{\code{get_measurements()}}{Get description of measurements served by this object}

\item{\code{toNEO4JDb(graph, colLabel = NULL)}}{Save an MRexperiment object to a Neo4j Graph database.}

\item{\code{toNEO4jDump()}}{Save an MRexperiment object to a Neo4j Graph database. The dump file is location at ~/dump.cypher.}

\item{\code{update(new_object, send_request = TRUE)}}{Update underlying data object with new object}
}}

