% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizMetagenomicsData-class.R
\docType{class}
\name{EpivizMetagenomicsData-class}
\alias{EpivizMetagenomicsData-class}
\alias{EpivizMetagenomicsData}
\title{Data container for MRexperiment objects}
\description{
Used to serve metagenomic data (used in e.g., icicle plots and heatmaps). Wraps
\code{\link[metagenomeSeq]{MRexperiment-class}} objects.
}
\section{Methods}{

\describe{
\item{\code{df_to_tree(root, df)}}{Helper function to recursively build nested response for getHierarchy

\describe{
  \item{root}{Root of subtree}
  \item{df}{data.frame containing children to process}
}
\value{root}{Nested tree structure}}

\item{\code{get_default_chart_type()}}{Get name of default chart type for this data type}

\item{\code{get_measurements()}}{Get description of measurements served by this object}

\item{\code{getAlphaDiversity(measurements = NULL, start = 1, end = 1000)}}{Compute alpha diversity using vegan for the given samples

\describe{
\item{measurements}{Samples to compute alpha diversity}
\item{start}{Start of feature range to query }
\item{end}{End of feature range to query}
}
\value{result}{List of alpha diversity values for given measurements}}

\item{\code{getCombined(measurements = NULL, seqName, start = 1, end = 1000,
  order = NULL, nodeSelection = NULL, selectedLevels = 3)}}{Return the counts aggregated to selected nodes for the given samples

\describe{
\item{measurements}{Samples to get counts for}
\item{seqName}{name of datasource}
\item{start}{Start of feature range to query}
\item{end}{End of feature range to query}
\item{order}{Ordering of nodes}
\item{nodeSelection}{Node-id and selectionType pairs}
\item{selectedLevels}{Current aggregation level}

}
\value{result}{List of samples with aggregate counts per feature}}

\item{\code{getHierarchy(nodeId = NULL)}}{Retrieve feature hierarchy information for subtree with specified root

\describe{
  \item{nodeId}{Feature identifier with level info}
}
\value{ret_data_frame}{List containing hierarchy of subtree}}

\item{\code{getPCA(measurements = NULL, start = 1, end = 1000)}}{Compute PCA over all features for given samples

\describe{
\item{measurements}{Samples to compute PCA over}
\item{start}{Start of feature range to query }
\item{end}{End of feature range to query}
}
\value{result}{List of PC1, PC2, and percent variance explained for each measurements}}

\item{\code{getRows(measurements = NULL, start = 1, end = 1000, selectedLevels = 3,
  selections = NULL)}}{Return the sample annotation and features within the specified range and level
    
    \describe{
      \item{measurements}{Samples to retrieve for}
      \item{start}{Start of feature range to query}
      \item{end}{End of feature range to query}
      \item{selections}{Node-id and selectionType pairs}
      \item{selectedLevels}{Current aggregation level}

    }
    \value{data_rows}{List of annotations for a given sample and features}
    }

\item{\code{getValues(measurements = NULL, start = 1, end = 1000,
  selectedLevels = 3, selections = NULL)}}{Return the counts for a sample within the specified range

\describe{
\item{measurements}{Samples to get counts for}
\item{start}{Start of feature range to query}
\item{end}{End of feature range to query}
\item{selections}{Node-id and selectionType pairs}
\item{selectedLevels}{Current aggregation level}

}
\value{data_columns}{List of counts for sample as selected level of hierarchy}}

\item{\code{propagateHierarchyChanges(selection = NULL, order = NULL,
  selectedLevels = NULL, request_with_labels = FALSE)}}{Update internal state for hierarchy

\describe{
  \item{selection}{Node-id and selectionType pairs}
  \item{order}{Ordering of features}
  \item{selectedLevels}{Current aggregation level}
  \item{request_with_labels}{For handling requests using fData entries from MRexperiment}
}}

\item{\code{row_to_dict(row)}}{Helper function to format each node entry for getHierarchy response

\describe{
  \item{row}{Information for current node.}
  }
\value{toRet}{Formatted named list.}}

\item{\code{searchTaxonomy(query = NULL, max_results = 15)}}{Find feature using text-based search

\describe{
\item{query}{String of feature for which to search}
\item{max_results}{Maximum results to return}

}
\value{result}{List of features that contain the substring query}}

\item{\code{update(new_object, send_request = TRUE)}}{Update underlying data object with new object}
}}

\examples{

library(metagenomeSeq)
data(mouseData)
obj <- metavizr:::EpivizMetagenomicsData$new(mouseData, feature_order = colnames(fData(mouseData)))

}
