% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizMetagenomicsData-class.R
\docType{class}
\name{EpivizMetagenomicsData-class}
\alias{EpivizMetagenomicsData}
\alias{EpivizMetagenomicsData-class}
\title{Data container for MRexperiment objects}
\description{
Used to serve metagenomic data (used in e.g., icicle plots and heatmaps). Wraps
\code{\link[metagenomeSeq]{MRexperiment-class}} objects.
}
\section{Methods}{

\describe{
\item{\code{get_default_chart_type()}}{Get name of default chart type for this data type}

\item{\code{get_measurements()}}{Get description of measurements served by this object}

\item{\code{toMySQLDb(con, colLabel = NULL)}}{Save an MRexperiment object to a MySQL database.}

\item{\code{toNEO4JDb(graph)}}{Save an MRexperiment object to a Neo4j Graph database.}

\item{\code{toNEO4JDbHTTP(batch_url, neo4juser, neo4jpass, datasource, colLabel = NULL)}}{Save an MRexperiment object to a Neo4j Graph database.}

\item{\code{toNEO4jDump(filePath = "~/", fileName = "neo4j_dump.cypher")}}{Save an MRexperiment object to a Neo4j Graph database. set the location of the output file with fileName and filePath.}

\item{\code{update(new_object, send_request = TRUE)}}{Update underlying data object with new object}
}}

