% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetavizApp-class.R
\docType{class}
\name{MetavizApp-class}
\alias{MetavizApp-class}
\alias{MetavizApp}
\title{Class managing connection to metaviz application.}
\description{
Class managing connection to metaviz application.
}

