% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetavizGraph-class.R
\docType{class}
\name{MetavizGraph-class}
\alias{MetavizGraph-class}
\alias{MetavizGraph}
\title{Graph implementation to query hierarchical feature data}
\description{
Used to manage aggregation and range queries from the Metaviz app UI.
}

