% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-methods.R
\docType{methods}
\name{register,phyloseq-method}
\alias{register,phyloseq-method}
\title{Generic method to register data to the epiviz data server}
\usage{
\S4method{register}{phyloseq}(object, type = "LeafCounts", ...)
}
\arguments{
\item{object}{The object to register to data server}

\item{...}{Additonal arguments passed to object constructors}

\item{columns}{Name of columns containing data to register}
}
\value{
An \code{\link{phyloseq-class}} object
}
\description{
Generic method to register data to the epiviz data server
}
