% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateMRExperiment.R
\name{validateObject}
\alias{validateObject}
\title{validate \code{\link[metagenomeSeq]{MRexperiment-class}} object}
\usage{
validateObject(object)
}
\arguments{
\item{an}{object of class \code{\link[metagenomeSeq]{MRexperiment-class}}}
}
\value{
TRUE or FALSE
}
\description{
validate \code{\link[metagenomeSeq]{MRexperiment-class}} object
}
